requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
config==0.5.1
distro==1.9.0
frozenlist==1.8.0
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.15.0
yarl==1.22.0
gspread==6.1.2
google-auth==2.33.0
pypdf==4.3.1
